Move Arch install functions to distributions.arch
